requirements.txt
Django==1.8.3
bootstrap-admin==0.3.6
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
wsgiref==0.1.2
django-daterange-filter==1.1.1
